backend/requirements.txt
# Backend 의존성 (Cloud Functions용)
firebase-functions
firebase-admin
google-generativeai
